+++ pareg_core/src/proc/err.rs
@@ -5,7 +5,7 @@ use quote::quote;
 use syn::LitStr;
 use thiserror::Error;
 
-pub type Result<T> = std::result::Result<T, Error>;
+pub type Result<T, E = Error> = std::result::Result<T, E>;
 
 #[derive(Debug, Error)]
 pub enum Error {