+++ pareg_core/src/proc/from_arg.rs
@@ -1,12 +1,11 @@
-use proc_macro2::{Literal, Span, TokenStream, TokenTree};
+use proc_macro2::{Literal, Span, TokenStream};
 use quote::{ToTokens, quote};
 use syn::{
     Attribute, Data, DeriveInput, Expr, Fields, Ident, LitChar, LitStr, Meta,
-    Token, parse::ParseStream, parse2, punctuated::Punctuated,
-    spanned::Spanned,
+    parse2, spanned::Spanned,
 };
 
-use crate::proc::{Error, Result};
+use crate::proc::{Error, Result, utils::extract_attribute_list};
 
 #[derive(Default)]
 struct FromArgConfig {
@@ -239,42 +238,4 @@ impl FromArgConfig {
     pub fn split(&self) -> char {
         self.split.unwrap_or('=')
     }
-}
-
-pub fn extract_attribute_list(
-    attribute: Attribute,
-) -> Result<Punctuated<TokenStream, Token![,]>> {
-    match attribute.meta {
-        Meta::List(l) => Ok(l.parse_args_with(punctuated_streams)?),
-        Meta::Path(_) => Ok(Punctuated::default()),
-        _ => {
-            Error::msg_span(attribute.span(), "Invalid attribute style.").err()
-        }
-    }
-}
-
-pub fn punctuated_streams(
-    input: ParseStream,
-) -> syn::Result<Punctuated<TokenStream, Token![,]>> {
-    let mut res = Punctuated::new();
-    loop {
-        let mut tokens = TokenStream::new();
-
-        while !input.is_empty() && !input.peek(Token![,]) {
-            tokens.extend([input.parse::<TokenTree>()?]);
-        }
-
-        if tokens.is_empty() && input.is_empty() {
-            break;
-        }
-
-        res.push_value(tokens);
-
-        let Ok(c) = input.parse() else {
-            break;
-        };
-        res.push_punct(c);
-    }
-
-    Ok(res)
 }